Broke tests out into own file. Clean up.
